+++ source/app/Http/Controllers/Api/CompanyController.php
@@ -22,7 +22,7 @@ class CompanyController extends Controller
     {
         $companies = Company::all();
         return response()->json([
-            'data'=>$companies
+            'data' => $companies
         ]);
     }
 
@@ -47,6 +47,9 @@ class CompanyController extends Controller
             'company_location' => 'nullable|string|max:255'
         ]);
 
+        $tempCompany = new Company($validate);
+        $mailService->sendCompanyAccountVerification($tempCompany, $logService);
+
         $validate['password'] = Hash::make($validate['password']);
 
         // Standaardwaarden voor ontbrekende velden
@@ -67,10 +70,10 @@ class CompanyController extends Controller
             }
         }
 
+        $this->authorize("create", Company::class);
         $company = Company::create($validate);
 
         $logService->setLog("Company", $company->id, "Company created", "Company", LogLevel::CRITICAL);
-        $mailService->sendCompanyAccountVerification($company, $logService);
 
         return response()->json([
             'data' => $company,
@@ -85,43 +88,9 @@ class CompanyController extends Controller
     {
         try {
             $company = Company::findOrFail($id);
-            return response()->json(['data'=>$company]);
-        } catch(ModelNotFoundException $e)
-        {
-            return response()->json(['message' => 'Company not found']);
-        }
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id, LogService $logService) : JsonResponse
-    {
-        try {
-            $company = Company::findOrFail($id);
-            $validated = $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|email|unique:companies,email,' .$company->id,
-            'password' => 'required|string|min:8',
-            'plan_type' => 'required|string|max:255',
-            'job_types' => 'required|string|max:255',
-            'job_domain' => 'required|string|max:255',
-            'booth_location' => 'required|string|max:255',
-            'photo' => 'nullable|string|max:255',
-            'speeddate_duration' => 'required|integer|max:60'
-            ]);
-
-            $validated['password'] = Hash::make($validated['password']);
-
-            $company->update($validated);
-            $logService->setLog("Company", $company->id, "Company updated", "Company", LogLevel::CRITICAL);
-
-            return response()->json([
-                'data' => $company,
-                'message' => 'Company updated successfully'
-            ]);
-        } catch (ModelNotFoundException $e)
-        {
+            $this->authorize("view", $company);
+            return response()->json(['data' => $company]);
+        } catch (ModelNotFoundException $e) {
             return response()->json(['message' => 'Company not found']);
         }
     }
@@ -133,6 +102,8 @@ class CompanyController extends Controller
     {
         try {
             $company = Company::findOrFail($id);
+
+            $this->authorize("delete", $company);
             $company->delete();
 
             $logService->setLog("Company", $company->id, "Company deleted", "Company", LogLevel::CRITICAL);
@@ -165,7 +136,7 @@ class CompanyController extends Controller
                 'job_requirements',
                 'job_description',
                 'company_location'
-                ];
+            ];
 
 
             // filtreren op enkel de informatie dat meegegeven wordt
@@ -178,6 +149,7 @@ class CompanyController extends Controller
 
             }
 
+            $this->authorize("update", $company);
             $company->update($data);
             return response()->json([
                 'data' => $company,
@@ -186,6 +158,40 @@ class CompanyController extends Controller
 
         } catch (ModelNotFoundException $e) {
             return response()->json(['message' => 'Company not found'], 404);
+        }
+    }
+
+    /**
+     * Update the specified resource in storage.
+     */
+    public function update(Request $request, string $id, LogService $logService): JsonResponse
+    {
+        try {
+            $company = Company::findOrFail($id);
+            $validated = $request->validate([
+                'name' => 'required|string|max:255',
+                'email' => 'required|email|unique:companies,email,' . $company->id,
+                'password' => 'required|string|min:8',
+                'plan_type' => 'required|string|max:255',
+                'job_types' => 'required|string|max:255',
+                'job_domain' => 'required|string|max:255',
+                'booth_location' => 'required|string|max:255',
+                'photo' => 'nullable|string|max:255',
+                'speeddate_duration' => 'required|integer|max:60'
+            ]);
+
+            $validated['password'] = Hash::make($validated['password']);
+
+            $this->authorize("update", $company);
+            $company->update($validated);
+            $logService->setLog("Company", $company->id, "Company updated", "Company", LogLevel::CRITICAL);
+
+            return response()->json([
+                'data' => $company,
+                'message' => 'Company updated successfully'
+            ]);
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['message' => 'Company not found']);
         }
     }
 }
